This fixes a major bug in the Argv parser. It also adds "true" and "false" as default aliases.
* AliasText.cs
DefaultAliasLoad: Sets "true" and "false" as default aliases.  
AddAlias: Removed an exception that is no longer required.  Overriding values is now allowed.  
* AliasTextTests.cs
GetDefaultAliases: Tests the new default alias code. 
GetBoolean: Changed this test so that it doesn't test using "true" and "false".  
* ArgvParser.cs: Updated this class with latest source code from the project page.  It is much more simple now.  
* ArgvConfigSourceTests.cs
AddSwitchCase: If a value is missing it does not return null.  It returns an empty string now.  
GetStringWithColon: Tests a problem with the old code and colons (':').